refactor admin dashboard screen components
